+++ index.html
@@ -10,7 +10,7 @@
 
 <body>
   <div id="app"></div>
-  <script type="module" src="./src/main2.js"></script>
+  <script type="module" src="./src/index.js"></script>
 </body>
 
 </html>
